Share social links across mobile apps
