+++ mobile/apps/auth/lib/ui/settings_page.dart
@@ -18,7 +18,6 @@ import 'package:ente_auth/ui/settings/general_settings_page.dart';
 import 'package:ente_auth/ui/settings/more_from_ente_section.dart';
 import 'package:ente_auth/ui/settings/notification_banner_widget.dart';
 import 'package:ente_auth/ui/settings/security_settings_page.dart';
-import 'package:ente_auth/ui/settings/social_icons_row.dart';
 import 'package:ente_auth/ui/settings/support_settings_page.dart';
 import 'package:ente_auth/ui/settings/theme_settings_page.dart';
 import 'package:ente_auth/utils/dialog_util.dart';
@@ -26,6 +25,7 @@ import 'package:ente_components/ente_components.dart';
 import 'package:ente_lock_screen/local_authentication_service.dart';
 import 'package:ente_strings/ente_strings.dart';
 import 'package:ente_ui/components/settings/app_engagement_section.dart';
+import 'package:ente_ui/components/settings/social_icons_row.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:hugeicons/hugeicons.dart';
